Port the remaining analyses from notebooks to scripts
Each of the four remaining `marimo` notebooks becomes a script of the same name in
`./scripts`, run by its rule through `script:`. The translation is cell by cell in file
order, which is the order the notebooks already ran in, so the analysis itself is
unchanged; what goes away is the machinery around it. The pickled context, the
subprocess that ran `marimo export html`, the string replacement done to its output,
and the `*_context.pickle` outputs of five rules are all gone, as is the stub-context
code each notebook carried for `marimo edit`.

`seqneut_funcs.py` holds what more than one script needs. `get_median_bound` was
copy-pasted verbatim into two notebooks, and the rounding and the Pearson correlation
of log titers were each written twice with small differences; there is now one of each.

Three latent problems in the notebooks do not survive the translation. The cell writing
`titers.csv` filtered on `viruses_to_drop` without declaring it, so nothing in
`marimo`'s dependency graph forced the cell defining it to run first; it worked only
because independent cells run in file order. The manual-drops loop had a second,
unreachable `elif` for `barcode_serum_replicates` that read a column existing only
inside the first branch. And its `barcode_wells` branch built a mask from the
unfiltered counts while applying it to the filtered ones, which pandas resolved by
reindexing, with a warning, and correctly only because the index happened to be
preserved.

Verified against a `marimo` run of the test example: all 571 output CSV and YAML files
byte-identical, and all 79 charts across the 16 reports with identical `vega-lite`
specifications once `altair`'s internal view names and `marimo`'s empty `usermeta` are
set aside. All three configurations run, and `test_titers_as_expected.py` passes for
both the default and the collapsed-barcode runs. The reports are two to four times
smaller and no longer fetch a front end from a CDN to display themselves.

`curve_display_method` loses `inline`, which meant handing the figure to `marimo`;
`neutcurve` rejects an invalid value when a figure is drawn, so the workflow does not
check it again. `curves.pdf` is written without the creation timestamp `matplotlib`
embeds, so it too is reproducible. `marimo` stays in
`environment.yml`, commented as unused by the pipeline, so that other analyses sharing
the environment keep it.
