+++ .github/workflows/test.yaml
@@ -42,10 +42,6 @@ jobs:
         shell: bash -el {0}
         run: snakefmt --check .
 
-      - name: check marimo notebooks
-        shell: bash -el {0}
-        run: marimo check --ignore-scripts .
-
       - name: test code with pytest
         shell: bash -el {0}
         run: pytest
