--- scripts/run_marimo_w_context_pickle.py
@@ -1,98 +1,0 @@
-"""Run a marimo notebook with snakemake context passed via pickle."""
-
-import os
-import pickle
-import subprocess
-import sys
-
-# Redirect stderr and stdout to log
-# `noqa: SIM115` as this log file must stay open for the life of the script
-log_file = open(snakemake.log[0], "w")  # noqa: SIM115
-sys.stderr = sys.stdout = log_file
-
-# Build context dictionary from snakemake object
-context = {
-    "workdir": os.getcwd(),
-    "input": dict(snakemake.input),
-    "output": dict(snakemake.output),
-    "params": snakemake.params,
-    "threads": snakemake.threads,
-    "resources": dict(snakemake.resources),
-    "wildcards": dict(snakemake.wildcards),
-}
-
-marimo_nb = snakemake.input.marimo_nb
-marimo_html = snakemake.output.marimo_html
-context_pickle = snakemake.output.context_pickle
-
-print(f"Running marimo notebook: {marimo_nb=}")
-print(f"Using context pickle: {context_pickle=}")
-print(f"Context keys: {list(context.keys())=}")
-for key, val in context.items():
-    if val:
-        print(f"  For {key}, context is: {val}")
-print(f"Output HTML: {marimo_html=}")
-
-# Write context to pickle file
-with open(context_pickle, "wb") as f:
-    pickle.dump(context, f, protocol=pickle.HIGHEST_PROTOCOL)
-
-# Run marimo export
-cmd = [
-    "marimo",
-    "export",
-    "html",
-    marimo_nb,
-    "-o",
-    marimo_html,
-    "--",
-    "--context-pickle",
-    context_pickle,
-]
-
-print(f"Running command: {' '.join(cmd)}")
-print("=" * 80)
-
-try:
-    # Stream output directly to log file (real-time logging)
-    result = subprocess.run(
-        cmd,
-        stdout=log_file,
-        stderr=subprocess.STDOUT,
-        check=False,
-    )
-    print("=" * 80)
-    if result.returncode != 0:
-        print(f"ERROR: marimo export failed with return code {result.returncode}")
-        raise subprocess.CalledProcessError(result.returncode, cmd)
-    print("marimo export completed successfully")
-except Exception as e:
-    print(f"ERROR: Exception running marimo export: {e}")
-    raise
-
-# Post-process HTML to set showAppCode to false
-print("Post-processing HTML to set showAppCode to false...")
-with open(marimo_html, "r") as f:
-    html_content = f.read()
-
-old_string = '"view": {"showAppCode": true}'
-new_string = '"view": {"showAppCode": false}'
-
-occurrence_count = html_content.count(old_string)
-if occurrence_count != 1:
-    error_msg = (
-        f"ERROR: Expected exactly 1 occurrence of '{old_string}' in HTML, "
-        f"but found {occurrence_count} occurrences"
-    )
-    print(error_msg)
-    raise ValueError(error_msg)
-
-html_content = html_content.replace(old_string, new_string)
-
-with open(marimo_html, "w") as f:
-    f.write(html_content)
-
-print("Successfully updated showAppCode setting in HTML")
-print(f"HTML output written to: {marimo_html}")
-
-log_file.close()
