+++ test_example/config.yml
@@ -24,7 +24,7 @@ viral_libraries:
 
 viral_strain_plot_order: data/viral_strain_plot_order.csv
 
-curve_display_method: svg  # {svg, pdf, png8, inline}; svg best resolution, png8 smallest
+curve_display_method: svg  # {svg, pdf, png8}; svg best resolution, png8 smallest
 
 neut_standard_sets:
   loes2023: data/neut_standard_sets/loes2023_neut_standards.csv
